// src/schemas/mock.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */


export interface paths {
  "/user/login": {
    /** 登录（手机号 + 密码） */
    post: {
      requestBody: {
        content: {
          "application/json": {
            /** @example 12345678901 */
            account: string;
            /** @example 123456 */
            password: string;
          };
        };
      };
      responses: {
        /** @description ok */
        200: {
          content: {
            "application/json": {
              /**
               * Format: int64
               * @example 0
               */
              code: number;
              /** @example ok */
              msg?: string;
              data?: {
                /** @example e82aceb7a991419492c58227d0a03a40 */
                accessToken?: string;
              };
            };
          };
        };
      };
    };
  };
  "/user/register": {
    /** 注册 */
    post: {
      requestBody: {
        content: {
          "application/json": {
            /** @example 12345678901 */
            phone: string;
            /** @example 小李 */
            nickname: string;
            /** @example 123456 */
            password: string;
            /** @example user@example.com */
            email?: string;
          };
        };
      };
      responses: {
        /** @description ok */
        200: {
          content: {
            "application/json": {
              /**
               * Format: int64
               * @example 0
               */
              code: number;
              /** @example ok */
              msg?: string;
              data?: {
                /** @example e82aceb7a991419492c58227d0a03a40 */
                accessToken?: string;
              };
            };
          };
        };
      };
    };
  };
  "/user/logout": {
    /** 退出 */
    post: {
      parameters: {
        header: {
          /** @example e82aceb7a991419492c58227d0a03a40 */
          api_key: string;
        };
      };
      responses: {
        /** @description ok */
        200: {
          content: {
            "application/json": {
              /**
               * Format: int64
               * @example 0
               */
              code: number;
              /** @example ok */
              msg?: string;
              /** @example null */
              data?: Record<string, never>;
            };
          };
        };
      };
    };
  };
  "/user/profile": {
    /** 获取用户信息 */
    post: {
      parameters: {
        header: {
          /** @example e82aceb7a991419492c58227d0a03a40 */
          api_key: string;
        };
      };
      responses: {
        /** @description ok */
        200: {
          content: {
            "application/json": {
              /**
               * Format: int64
               * @example 0
               */
              code: number;
              /** @example ok */
              msg?: string;
              data?: {
                /**
                 * Format: int64
                 * @example 1234567890
                 */
                uid?: number;
                /** @example evan */
                username?: string;
                /** @example 小李 */
                nickname?: string;
                /** @example 1 */
                gender?: number;
                /** @example http://localhost:3000/static/api/avatar.jpg */
                avatar?: string;
              };
            };
          };
        };
      };
    };
  };
}

export type webhooks = Record<string, never>;

export type components = Record<string, never>;

export type $defs = Record<string, never>;

export type external = Record<string, never>;

export type operations = Record<string, never>;